# docs/api/library-v1.openapi.yaml
openapi: 3.1.0
...
info:
  title: XiaDown Library API
  version: 1.0.0
  description: |
    Version 1 of the device-facing XiaDown Library API. The API is exposed only
    by a user-enabled Library transport. Pairing credentials and Bearer tokens
    are secrets; clients must not put them in URLs, analytics, logs, or crash
    reports.

    The pairing response grants `library.read` and `tasks.read` by default.
    `music.read`, `music.state`, `music.manage`, `rss.read`, `rss.state`,
    `rss.manage`, `rss.fetch`, `tasks.create`, and `tasks.control` can be granted
    later by the desktop owner. Pairing never grants either RSS management or
    public-feed fetch access by default.
    A grant is bound to one Catalog, so a valid token from another Catalog is
    rejected.

    A stable device ID and public-key hash are client assertions, not authority
    to inherit an existing grant. Pairing rejects collisions with an active
    grant. After the Desktop owner revokes a grant, a fresh short-lived pairing
    session may replace it in place: the grant identity and creation time remain
    for namespace continuity, but the token and public-key hash rotate and scopes
    reset to the default minimum. The old token remains invalid and prior managed
    scopes must be granted again by the Desktop owner.

    Canonical response property names are case-sensitive camelCase. Path and
    query parameter names are case-sensitive. Unless an input is explicitly
...
servers:
...
    variables:
...
security: []
...
paths:
...
  /api/v1/library:
    get:
      tags: [Library]
      operationId: getLibraryOverview
      summary: Read the default Catalog overview and synchronization cursor
      security:
        - bearerAuth: []
      x-xiadown-required-scope: library.read
      responses:
        "200":
          description: Current Catalog counts, health counts, and a high-water synchronization cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/tasks:
    get:
      tags: [Tasks]
      operationId: listTasks
      summary: List safe task projections
      security:
        - bearerAuth: []
      x-xiadown-required-scope: tasks.read
      parameters:
        - $ref: "#/components/parameters/TaskStatuses"
        - $ref: "#/components/parameters/TaskKinds"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/TaskLimit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A bounded task list. Local paths, raw URLs, logs, and process diagnostics are never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
        "400":
          description: Invalid pagination (`invalid_pagination`) or search query (`invalid_request`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Create a downloader task
      security:
        - bearerAuth: []
      x-xiadown-required-scope: tasks.create
      requestBody:
        required: true
        x-max-body-bytes: 32768
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task accepted by the Library application service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Malformed JSON, unknown fields, unsafe/private URL, unsupported mode/quality, or another invalid request (`invalid_request`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tasks/{id}:
    get:
      tags: [Tasks]
      operationId: getTask
      summary: Read one safe task projection
      security:
        - bearerAuth: []
      x-xiadown-required-scope: tasks.read
      parameters:
        - $ref: "#/components/parameters/OpaqueID"
      responses:
        "200":
          description: Current task state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tasks/{id}/cancel:
    post:
      tags: [Tasks]
      operationId: cancelTask
      summary: Ask the Library scheduler to cancel a task
      security:
        - bearerAuth: []
      x-xiadown-required-scope: tasks.control
      parameters:
        - $ref: "#/components/parameters/OpaqueID"
      responses:
        "200":
          description: Updated task state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/TaskStateConflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tasks/{id}/resume:
    post:
      tags: [Tasks]
      operationId: resumeTask
      summary: Ask the Library scheduler to resume a task
      security:
        - bearerAuth: []
      x-xiadown-required-scope: tasks.control
      parameters:
        - $ref: "#/components/parameters/OpaqueID"
      responses:
        "200":
          description: Updated task state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/TaskStateConflict"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: xd1 opaque device token
      description: "Send `Authorization: Bearer <token>`. The token is Catalog-bound and must be stored as a secret."
...
  parameters:
...
    OpaqueID:
      name: id
      in: path
      required: true
      description: Case-sensitive opaque identifier. Do not infer filesystem paths from IDs.
      schema:
        type: string
        minLength: 1
        maxLength: 255
    SearchQuery:
      name: q
      in: query
      description: Case-insensitive substring search. After percent-decoding, the value must be valid UTF-8 and at most 512 Unicode code points. Length is evaluated before endpoint-specific whitespace handling; overlong values are rejected, never truncated.
      schema:
        type: string
        maxLength: 512
...
    TaskLimit:
      name: limit
      in: query
      description: Task page size. Omitted or zero selects the default.
      schema:
        type: integer
        minimum: 0
        maximum: 200
        default: 100
    Offset:
      name: offset
      in: query
      description: Zero-based row offset.
      schema:
        type: integer
        minimum: 0
        default: 0
...
    TaskStatuses:
      name: status
      in: query
      description: Repeat to select multiple statuses. Matching is case-insensitive; unknown values yield no matches.
      x-case-insensitive: true
      style: form
      explode: true
      schema:
        type: array
        items:
          $ref: "#/components/schemas/TaskStatus"
    TaskKinds:
      name: kind
      in: query
      description: Repeat to select multiple task kinds. Matching is case-insensitive; unknown values yield no matches.
      x-case-insensitive: true
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request (`invalid_request`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Unauthorized:
      description: Missing, malformed, expired, revoked, or invalid Bearer token (`unauthorized`).
      headers:
        WWW-Authenticate:
          description: Bearer challenge for the XiaDown Library realm.
          schema:
            type: string
            const: 'Bearer realm="xiadown-library"'
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: The token belongs to another Catalog (`catalog_access_denied`) or lacks the operation scope (`insufficient_scope`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested public resource does not exist (`not_found`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    TaskStateConflict:
      description: The requested action is not valid in the current task state (`task_state_conflict`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Opaque server failure (`internal_error`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorCode:
      type: string
      description: Stable machine-readable API error code. Error messages and implementation details are intentionally absent.
      enum:
        - asset_stream_limit_reached
        - asset_unavailable
        - catalog_access_denied
        - insufficient_scope
        - idempotency_conflict
        - internal_error
        - invalid_cursor
        - invalid_music_mutation
        - invalid_pagination
        - invalid_request
        - music_content_changed
        - music_dependency_pending
        - music_entity_not_found
        - music_idempotency_conflict
        - music_revision_conflict
        - music_stream_limit_reached
        - not_found
        - pairing_failed
        - pairing_invalid
        - reset_required
        - rss_state_conflict
        - task_state_conflict
        - unauthorized
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorCode"
...
    CatalogStatus:
      type: string
      enum: [active, archived]
...
    TaskStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled]
    TaskMode:
      type: string
      x-case-insensitive: true
      enum: [quick, custom]
    TaskQuality:
      type: string
      x-case-insensitive: true
      enum: [best, bitrate, audio]
...
    Catalog:
      type: object
      additionalProperties: false
      required: [id, name, status, isDefault, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/CatalogStatus"
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryCounts:
      type: object
      additionalProperties: false
      required: [all, video, audio, books, images, others]
      properties:
        all:
          type: integer
          minimum: 0
        video:
          type: integer
          minimum: 0
        audio:
          type: integer
          minimum: 0
        books:
          type: integer
          minimum: 0
        images:
          type: integer
          minimum: 0
        others:
          type: integer
          minimum: 0
    StatusCounts:
      type: object
      additionalProperties: false
      required: [active, needsReview, missing, trashed]
      properties:
        active:
          type: integer
          minimum: 0
        needsReview:
          type: integer
          minimum: 0
        missing:
          type: integer
          minimum: 0
        trashed:
          type: integer
          minimum: 0
    HealthCounts:
      type: object
      additionalProperties: false
      required:
        - assetLinks
        - itemsWithoutAssets
        - unavailableAssetFiles
        - legacyFilesWithErrors
        - offlineStorageRoots
        - readOnlyStorageRoots
        - storageRootsWithErrors
      properties:
        assetLinks:
          type: integer
          minimum: 0
        itemsWithoutAssets:
          type: integer
          minimum: 0
        unavailableAssetFiles:
          type: integer
          minimum: 0
        legacyFilesWithErrors:
          type: integer
          minimum: 0
        offlineStorageRoots:
          type: integer
          minimum: 0
        readOnlyStorageRoots:
          type: integer
          minimum: 0
        storageRootsWithErrors:
          type: integer
          minimum: 0
    SyncState:
      type: object
      additionalProperties: false
      required: [epoch, cursor]
      properties:
        epoch:
          type: string
        cursor:
          type: integer
          format: int64
          minimum: 0
    LibraryOverview:
      type: object
      additionalProperties: false
      required: [catalog, categories, statuses, health, sync, capabilities]
      properties:
        catalog:
          $ref: "#/components/schemas/Catalog"
        categories:
          $ref: "#/components/schemas/CategoryCounts"
        statuses:
          $ref: "#/components/schemas/StatusCounts"
        health:
          $ref: "#/components/schemas/HealthCounts"
        sync:
          $ref: "#/components/schemas/SyncState"
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [snapshot-keyset-v1]
...
    TaskProgress:
      type: object
      additionalProperties: false
      properties:
        stage:
          type: string
        percent:
          type: integer
        current:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
        speed:
          type: string
        speedMetric:
          $ref: "#/components/schemas/TaskSpeedMetric"
        updatedAt:
          type: string
          format: date-time
    TaskMetrics:
      type: object
      additionalProperties: false
      required: [fileCount]
      properties:
        fileCount:
          type: integer
          minimum: 0
        totalSizeBytes:
          type: integer
          format: int64
          minimum: 0
        durationMs:
          type: integer
          format: int64
          minimum: 0
    Task:
      type: object
      additionalProperties: false
      required: [id, name, kind, status, metrics, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
          maxLength: 160
          description: Sanitized display name; source URLs, paths, credentials, and token-like values are replaced by a generic name.
        kind:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        domain:
          type: string
        platform:
          type: string
        uploader:
          type: string
        publishTime:
          type: string
        progress:
          $ref: "#/components/schemas/TaskProgress"
        metrics:
          $ref: "#/components/schemas/TaskMetrics"
        errorCode:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
    TaskList:
      type: object
      additionalProperties: false
      required: [tasks]
      properties:
        tasks:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/Task"
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required: [url]
      properties:
        url:
          type: string
          format: uri
          minLength: 1
          maxLength: 8192
          pattern: "^https?://"
          description: Public HTTP(S) source URL. User information, local/private/special-use destinations, and NUL characters are rejected.
        title:
          type: string
          maxLength: 512
        mode:
          $ref: "#/components/schemas/TaskMode"
        quality:
          $ref: "#/components/schemas/TaskQuality"
        formatId:
          type: string
          maxLength: 256
        audioFormatId:
          type: string
          maxLength: 256
        writeThumbnail:
          type: boolean
        subtitleLangs:
          type: array
          maxItems: 32
          items:
            type: string
            minLength: 1
            maxLength: 64
        subtitleAuto:
          type: boolean
        subtitleAll:
          type: boolean
        subtitleFormat:
          type: string
          maxLength: 128
        transcodePresetId:
          type: string
          maxLength: 256
        deleteSourceFileAfterTranscode:
          type: boolean